Add value model. See #26.
