--- files/rsum/rsum/home/__init__.py
@@ -1,1 +1,0 @@
-import models
